fix(predictioneval): carry the four 64-bit wire fields losslessly
The exported model carries four exact 64-bit patterns across a wire:
SuppliedDrawTrace.Words, OrderedRulesSelection.ShareBits,
OrderedRulesTraceEntry.ShareBits and OrderedRulesTraceEntry.RawWordValue.
They travelled as JSON numbers, and a JSON number cannot carry one: an
IEEE-754 consumer rounds anything past 2^53, so two words the strict
Bernoulli comparison distinguishes could arrive equal, and two shares one
ulp apart collapsed into one. ShareBits was the sharper case — it is
math.Float64bits of a share, and the bit pattern of any normal double is
past that line, so every value of it was affected rather than only the
large ones.

Introduce OrderedRulesHex64: a uint64 in memory, a canonical fixed-width
sixteen-character lower-case hex string on the wire. MarshalText calls
hex64, the function the digests have used since they were written, rather
than restating the format, so the wire and the digests cannot drift.
UnmarshalText is that function's exact inverse and refuses everything
else — wrong width, upper case, a 0x prefix, a sign, any byte that is not
a lower-case hex digit — and leaves the destination untouched when it
refuses. No decimal compatibility shim is needed or present:
encoding/json will not hand a JSON number to a TextUnmarshaler at all, so
the old representation is closed off by the type.

All arithmetic stays on uint64 and every conversion is a reinterpretation
of the same bits, so the mechanism semantics are unchanged and none of
the four digests moved. No version constant is bumped: the semantics did
not change, only the transport.

Falsifiers cover every boundary the audit named — zero, leading-zero
values, both sides of 2^53, MaxUint64, two shares one ulp apart, and the
bit pattern of 1e-300 — in both directions, plus the refusal table, the
four fields in their real carriers, and a type-agnostic check that no
number anywhere in an encoded evaluation is past 2^53. The loss is
demonstrated rather than argued: the old and new representations are read
through the same IEEE-754 consumer and the first is shown collapsing four
pairs of distinguishable values while the second keeps them apart. The
digests are pinned against values captured by running the same fixture on
a3187b2, the commit before this change.

Two behaviours are recorded as NOT repaired because they are outside this
change: a sixteen-character all-decimal spelling is a well-formed hex
word and is read as its hexadecimal value, and a JSON null remains a
no-op rather than an error, exactly as it was when the fields were plain
uint64.

+++ internal/predictioneval/ordered_rules_bindings_test.go
@@ -248,12 +248,13 @@ func TestOrderedRulesAllZeroPoolProducesZeroSharesNotNaN(t *testing.T) {
 	if sel.OutcomeIdentity != "A" {
 		t.Fatalf("selected %q, want \"A\"", sel.OutcomeIdentity)
 	}
-	if math.IsNaN(math.Float64frombits(sel.ShareBits)) || sel.ShareBits != math.Float64bits(0) {
+	if math.IsNaN(math.Float64frombits(uint64(sel.ShareBits))) ||
+		uint64(sel.ShareBits) != math.Float64bits(0) {
 		t.Fatalf("the recorded share is %#016x, want an exact zero; an all-zero pool must not divide "+
-			"zero by zero", sel.ShareBits)
+			"zero by zero", uint64(sel.ShareBits))
 	}
 	for _, e := range ev.Trace {
-		if math.IsNaN(math.Float64frombits(e.ShareBits)) {
+		if math.IsNaN(math.Float64frombits(uint64(e.ShareBits))) {
 			t.Fatalf("a NaN share reached the trace at outcome %d", e.OutcomeIndex)
 		}
 	}
